Weighted-time difficulty adjustment
To facilitate faster continuous retargeting, two weights are
applied to the individual inter-block times within a sliding
adjustment window.

The first weight is a multiplicative adjustment for the difficulty
as-of the block itself, relative to the difficulty of the last
block in the window.

The second weight is a simple arithmetic weight decreasing in
equal steps from the most recent block in the window, down to
zero for the block just prior to the window.